Power_Num file modified

+++ Power_Num.cpp
@@ -1,7 +1,7 @@
 #include <iostream>
 using namespace std;
 int power(int base,int expo){
-    double ans =1;
+    int ans =1;
     // This is Function to find power by calling function without using header file
     for(int i=1;i<=expo;i++){
         ans*=base;
